refactor: Refactored Routes Files
Refactored routes files for ease of testing
