+++ src/routes/JournalsRoute/JournalsRoute.js
@@ -5,9 +5,9 @@ import { FormattedMessage } from 'react-intl';
 import { AppIcon } from '@folio/stripes/core';
 
 import { SASQRoute } from '@k-int/stripes-kint-components';
-import { OAFilterHeaderComponent } from '../components/SearchAndFilter';
-import { findIssnByNamespace } from '../util/journalUtils';
-import Journal from '../components/views/Journal';
+import { OAFilterHeaderComponent } from '../../components/SearchAndFilter';
+import { findIssnByNamespace } from '../../util/journalUtils';
+import Journal from '../../components/views/Journal';
 
 const JournalsRoute = ({ path }) => {
   const renderISSN = (d, namespace) => {
